Initial import of osel code
This is an initial import of the osel codebase.  The osel tool is a tool that
initiates external security scans (initially through Qualys) upon reciept of
AMQP events that indicate certain sensitive events have occurred, like a
security group rule change.

The commit history had to be thrown away because it contained some non-public
data, so I would like to call out the following contributors:

- Charles Bitter https://github.com/cbitter78
- Olivier Gagnon <Olivier_Gagnon@cable.comcast.com>
- Joseph Sleiman <Joseph_Sleiman@comcast.com>

This uses go 1.10 and vgo for dependency management.

Change-Id: Ib6abe2024fd91978b783ceee4cff8bb4678d7b15